requirements.txt
# LLM (install the one you use)
openai>=1.30.0
anthropic>=0.25.0

# Config
pyyaml>=6.0.1

# Flask helper (also used in app/requirements.txt for the Docker image)
flask>=3.0
pymysql>=1.1
gunicorn>=21.0

# Streamlit UI
streamlit>=1.35.0

# Dev / testing
pytest>=8.0.0
